airdrop-farming-bot 2/requirements.txt
web3==6.20.1
eth-account>=0.9.0
pydantic==2.8.2
pyyaml==6.0.2
rich==13.7.1
tenacity==8.5.0
python-dotenv==1.0.1
eth-abi>=4.2.1
eth-utils>=2.3.1
